+++ .changeset/shy-pans-report.md
@@ -1,0 +1,5 @@
+---
+'@leafygreen-ui/mongo-nav': minor
+---
+
+Adds client-side filetring to OrganizationSelect and ProjectSelect
