+++ packages/leafygreen-provider/src/index.ts
@@ -1,6 +1,6 @@
-import LeafyGreenProvider from './LeafyGreenProvider';
-import { useUsingKeyboardContext } from './UsingKeyboardProvider';
-import { useBaseFontSize } from './TypographyProvider';
+import LeafyGreenProvider from './LeafyGreenContext';
+import { useUsingKeyboardContext } from './UsingKeyboardContext';
+import { useBaseFontSize } from './TypographyContext';
 
 export { useUsingKeyboardContext, useBaseFontSize };
 export default LeafyGreenProvider;
